requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pandas==2.2.3
pytest==8.3.4
pytest-cov==6.0.0
numpy==2.0.2
scikit-learn==1.6.1
uvicorn==0.34.0
granian==2.2.6
pydantic-settings==2.9.1
sqlmodel==0.0.24
asyncpg==0.30.0
matplotlib==3.10.1
alembic==1.15.2